requirements.txt
groq==0.30.0
python-dotenv==1.1.1
python-telegram-bot==22.3
